requirements.txt
Flask
numpy
tensorflow
pandas
joblib
pymongo
python-dotenv
